requirements.txt
kafka-python==2.0.2
pyspark==3.4.0
psycopg2-binary==2.9.7
pandas==2.0.3
numpy==1.24.3